Update patch set 1
Patch Set 1:

I added support only for the "offload" options and not the extra_options because pre/post-up/down are not implemented to follow as a model, and not sure that my basic skills in Perl would be enough to pull a decent implementation.  Also, being a multiplicity of items makes them harder to manage (set and unset all at once?  append only?  remove individually? ...), considerably more care is needed compared to the basic options.

I suppose that the management of this option is not-so-frequent and sufficiently complex that it's preferable to edit them directly on network.yml?  That would be my option of choice when needing to add them, instead of learning how to do the proper invocation of ngcp-network every time.  (I suppose that this is also the reason why pre/post-up/down rules, only pre-up until recent times, were not managed by this tool).

Patch-set: 1
Label: Code-Review=0