TT#7570 Add /usr to the list of directories to manage configs for
The monit-services OCF Resource Agent for pacemaker needs to be
installed there.

Ideally we would not need this, but as this is the only location
allowed by pacemaker, let's add it, even if we might at some point
end up not using it, as we can just not install any templates in those
directories.

Change-Id: Ib14b68388f018b45bc4d7cf6895ca17d0fc802a3