TT#62401 Special case ngcp-service files and generate them always first
These files are an intermediate data source, which are being used also
during builds of templates. We cannot make them first stage YAML sources
as they do depend on information coming from the first stage YAML
sources themselves. So we need to special-case them to make sure they
are always built first, so that the other files have up-to-date
information to use when being built.

Change-Id: I6cfd6e80823ddfaf71e2246de5518b9cb6d6a1d2