TT#63758 Add 'ngcpcfg message' to provide better logging possibilities
'ngcpcfg push-parallel' on Carrier requires the next level of debug
messages readability to prevent users confusing in both console and logs.

At the same time we should be strict in console (as we normally apply 10+
nodes in parallel) and be verbose in logs to provide users with all
the details about the failed ngcpcfg call on the particular node.

The new 'ngcpcfg message' tool allows us to write something into logs
for remote Carrier node on 'ngcpcfg push-parallel' using solid logs style
and verbosity.

Change-Id: Ie9a81f9a4e50469680e5eb8370f45707ba8e0119