TT#63758 Fix lib/get_ngcp_neighbours on Carrier. Use get_hostname() instead of get_nodename()
lib/get_ngcp_neighbours generated /etc/default/ngcp-roles which contains
two lists "NGCP_HOSTS" and "NGCP_NEIGHBOURS". The first list is "all nodes
in carrier cluster", while the second is "all nodes in carrier cluster
exclusing this node".

On PRO (sp1 node):
> NGCP_HOSTS="sp1 sp2"
> NGCP_NEIGHBOURS="sp2"

On Carrier (web01a node):
> NGCP_HOSTS="db01a db01b lb01a lb01b prx01a prx01b web01a web01b"
> NGCP_NEIGHBOURS="db01a db01b lb01a lb01b prx01a prx01b web01b"

The problem was in get_nodename which is always sp1/sp2 on PRO AND Carrier.
As a result the code worked on PRO but didn't exclude the current node on Carrier.

This is a fix to exclude current node/hostname from neighbours list.

Change-Id: Icef4c3811c7e1b087157f57f9cb723b5a9456943