TT#15857 Fix DBI errors output
The DBI connect call can be specified the PrintError option, which will
make the function print the contents of the $DBI:errstr, but then we do
that again, which results in redudant output, so we disable this option.

The $DBI::errstr variable contains an error string that is *not*
terminated with a newline. The perl die built-in has magic behavior
when the string it gets does not end in a newline character, it will
then emit the source line, file name and a newline where the error
happened, which is in most cases not what we want. We just avoid
this by adding an explicit newline.

We also replace a print + exit with a die, and a printf without a
terminating newline which would mess up further output with a warn
call with an explicit newline, which will also make it go to stderr.

Change-Id: I348b8cc676da7f255c6c823de2a774694dbca826