Update patch set 2
Patch Set 2:

> Of course another solution would be to support multiple api_int
 > listening sockets :)

Sure, but I think I got lost with the back and forth in the task, as to what that might imply. For now making the check match reality is probably more important, and we can always lift it once this changes. :)

Patch-set: 2
Reviewer: Gerrit User 1000016 <1000016@85315e57-5f03-49c3-83f8-201ae2313a75>