TT#71952 Parallelize templated output generation
Switch away from the old daemon/wrapper architecture, which complicates
things and requires more scaffolding code.

We turn the daemon/wrapper and build_config into a single perl process
that will load all YAML files, and then process each input/output pair
on a parallel child, to try to speed up the processing as much as
possible.

For whole rebuilds, it might speed up the generation by at least a
factor of x2, x4 or more, depending on the number of active processors.

Change-Id: I51aa2f90336e34a20983d8733f45b64d9b6fea0b