Update patch set 8
Patch Set 8:

> Negative effect: on VMs LA is ~14, so maybe we have smarted code
 > here, to create amount of working threads 2*CPUs only, or something
 > smarter. Otherwise Vagrants/low_performance_VMs will be struggling
 > due to DOS on OS.

This would require a process pool implementation. I'd say let's merge this and see how it goes and we can always check how to improve further if it causes problems. Also consider that a spike in the load should be fine as that's a very bounded execution.

Patch-set: 8
Reviewer: Gerrit User 1000016 <1000016@85315e57-5f03-49c3-83f8-201ae2313a75>