Update patch set 5
Patch Set 5:

> This is really strange.
> Right now I do not see any performance improvements here,
> moreover new version is currently 0.1-0.2 seconds slower in my tests for 16 threads on 16 CPU.
> 
> Having such tests results I believe it is better to keep the old version, isn't it?
> The old version is easier to read and it is faster for 1/2/4/8/16 threads I tested.

As I understand it, the main difference is that the chmod was moved from every re-run, to just the time when the file is created.

The difference probably be very small, almost negligible, a few milliseconds at most, specially if files/inodes are already in the fs cache and there's no commit to disk in between.

So I think that the difference is supposed to come, again AFAIUI, from "doing the right thing" and paying the price once by setting the permissions when the file is created, instead of having the recurrent payment every time that the file is revisited.  But it's more of a conceptual benefit than a material/measurable one, I think.

That said, for me it's valid if you object to changing the code because of the lack of improvement and code more difficult to read.  I am not very decided one way or the other.

Patch-set: 5
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>