TT#79550 allow interfaces with just shared IP
Add a new config flag to interfaces to allow usage with just a shared IP
address configured, without a static IP address. This requires special
routing to be set up by the operator and can be used on dummy interfaces
as well.

We explicitly disallow this usage for any interface that has a type of
ha_int or ssh_* or mon_*.

Change-Id: Id950a525d2f499fdd17a0a9fee876f01e9f01d87