TT#81050 Add HA_CONFIG and PAIR_CONFIG support
This makes it possible to use more specific config.yml files either for
a pair of nodes, for settings that will always affect both, or for the
HA names, for settings that might affect all A or B nodes.

This matches the support in the template files.

We need to move sourcing the ha_features and carrier_features before
souring init, as now the ngcpcfg.cfg file can use these variables if
set.

Change-Id: Ic6b32acceb8f020f48c3afa09d9e84252b3531e7