TT#80164 Check permissions and ownership of customtt and patchtt files
Having different permissions or ownership can be dangerous, lead to errors and
wasting time finding problems, specially at critical times like emergency fixes
or upgrades.

So check that all the customtt and patchtt files have the same ownership and
permissions as their base files, or otherwise bark and stop.

Check also that the base file exist, otherwise complain (it might be checked
elsewhere, but it's important to do it in this function as well).

Example, given the files:

  root@spce:~# ls -ld /etc/[abcd]{,.*tt}
  -rw-r--r-- 1 daemon daemon 0 Jun  8 18:24 /etc/a
  -rw-r--r-- 1 root   root   0 Jun  8 18:05 /etc/a.customtt
  -rw------- 1 root   root   0 Jun  8 19:48 /etc/b
  -rw-r--r-- 1 root   root   0 Jun  8 19:48 /etc/b.patchtt
  -rw-r--r-- 1 root   root   0 Jun  8 19:48 /etc/c.patchtt
  -rw-r--r-- 1 root   root   0 Jun  8 19:55 /etc/d.customtt

this is the result of the execution:

  root@spce:~# ngcpcfg check
  2020-06-08 19:58:08 spce: yml configs were validated successfully
  2020-06-08 19:58:08 spce: Error: Permissions of customtt/patchtt file and base file differ
  -rw-r--r-- 1 daemon daemon 0 Jun  8 18:24 /etc/a
  -rw-r--r-- 1 root   root   0 Jun  8 18:05 /etc/a.customtt
  2020-06-08 19:58:08 spce: Error: Base file corresponding to customtt/patchtt file '/etc/c.patchtt' not found
  2020-06-08 19:58:08 spce: Error: Permissions of customtt/patchtt file and base file differ
  -rw------- 1 root root 0 Jun  8 19:48 /etc/b
  -rw-r--r-- 1 root root 0 Jun  8 19:48 /etc/b.patchtt
  2020-06-08 19:58:08 spce: Error: Base file corresponding to customtt/patchtt file '/etc/d.customtt' not found

Added unit tests to ensure that this works as intended, although they are a bit
limited because permissions cannot be changed as freely as in normal systems,
and when adding tests with names "patchtt" it interferes with other tests in the
suite, so using only names containing "customtt".

Change-Id: I19a8446e33ce51b0796af3e0c2c85cc2fc73caaa