Update patch set 2
Patch Set 2:

> Since you choose this way,

I have to say that I didn't put much thought about it, I was merely following your suggested ideas at the ticket, https://sipwise.my.workfront.com/task/view?ID=5e9ee23901555e50926566bf6b160a7b , which was pretty much what I implemented (I think?).

Thinking about it a bit more, I am not huge fan of the idea of doing any sort of modification in operations named "check", even if it's relatively straightforward and it's supposed to make things saner -- for me RW actions on a RO-concept like "check" violates the "principle of least surprise".

Maybe I can extend the implementation of "ngcpcfg check" to not modify directly, but *ask*, if it should fix some of the issues when they are found -- so it doesn't violate this principle :)

> you have to create an upgrade step to check perms before the upgrade. Isn't it? Otherwise during upgrade mr8.4->mr8.5 ngcp-upgrade will be aborted due to wrong perms.

Since it's the first time that we run it, mr8.4->mr8.5 (or more likely mr7.5->mr8.5) will find more of these cases, but it can happen in the future (a customtt can be created at any point with the wrong permissions), so I think that it's a general problem?

So if we do something like that it should be a permanent step, IMO?

I think that leaving the steps as they are, and implementing the feature above of "ask to fix the permissions if a problem is found" as an optional action of "ngcpcfg check", backporting it to mr7.5+, is all that we need to do?

> IMHO, I would just force patchtt/customtt perms to be equal to tt2 files.
> So, ngcpcfg will set them identically instead of producing the noise here.
> What do you think?

I replied above, but to comment on this specifically, I think that it's relatively safe to let the upgrade abort at that point for the Ops to take corrective action -- this happens in the "pre-check phase" and no modifications important modifications to the system have been made at that point, so it's as if we abort for other reasons (node not on stand-by, problem accessing apt repos even if only because of network blips, problems validating YML files that "ngcpcfg check" also looks for, many others).

More in general, some things we can fix, e.g. permissions and ownership, but if there's a customtt/patchtt file and the base file does not exist, which I already implemented, it's not clear to me what action should be taken, and perhaps we implement more of these in the future.

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>