Update patch set 2
Patch Set 2:

> > Thinking about it a bit more, I am not huge fan of the idea of doing any sort of modification in operations named "check", even if it's relatively straightforward and it's supposed to make things saner -- for me RW actions on a RO-concept like "check" violates the "principle of least surprise".
> 
> I do not want to perform RW in RO "check", but we can fix patchtt perms in "ngcpcfg patch" which is a part of "ngcpcfg build". As for perms for "customtt" maybe inside "build" or maybe inside "patchtt" too to keep all those code in one place.

Sigh, I lost a comment once, let's try again :(

OK, I misunderstood you, but then I see a problem, more on this below.

> So if I need to have customtt/patchtt and didn't care enough about perms -> ngcpcfg can fix perms.
> 
> 
> The question is: 
> 1) should ngcpcfg change perms on customtt file
>  OR
> 2) ngcpcfg should build ${desination} file from ${desination}.customtt.tt2 but set perms from ${desination}.tt2 ?
> 
> IMHO 2) is the way to go. No need to ask someone or create upgrade steps at all. Also we will not change something on HDD, we will just use proper source for perms on "build" stage.

I think that doing 2, or even 2+1 (using the main file for permissions, but also setting the correct permissions for customtt/patchtt files), is good.  But that will work only for future systems.

For all existing systems, I think that we should add a check like this ngcp-upgrade, to avoid running into the same problem as Yaroslav already did once.

There's a small risk to stop the upgrade for cases in which it might not matter (or put it in pause and ask for confirmation if we want to continue), but also there's the benefit of pointing to the right problem and how it can be fixed there, instead of having to spend minutes/hours debugging problems like Yaroslav did, in an stressful situation and a limited period of a maintenance window.

I think that it would be also good to keep this check that I implemented here permanently, for extra things like customtt/patchtt files with missing "base file", and well, to enforce a situation that we don't want to happen -- it fits in the concept of "ngcpcfg check", IMO.

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>