Update patch set 2
Patch Set 2: Code-Review+1

The change itself looks good to me, but I don't understand fully the implications of the changes and outputs of commands and so on, so just +1.  But I'm OK if you merge it right now with the other +2 votes.

> Volodymyr is on top of mr8.1.1, what about those? Any further releases we still care about and aren't listed in supported ones?

AFAIK he's working on a special PPA for Magenta, probably they can cherry-pick and put into the PPA themselves (if notified appropriately when this one is merged).

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+1