Update patch set 2
Patch Set 2: Code-Review+2

> Then my wording wasn't good enough.
> 
> If your working directory is /etc and you invoke `ngcpcfg commit`, then the initial `git rev-parse --git-dir` from the `if ! git rev-parse --git-dir >/dev/null 2>&1 ; then` does NOT fail, so we are not executing the following `cd "$(dirname "${0}")"` then, while we expect to always run it from inside /etc/ngcp-config, as that's where we store and operate with the .ngcpcfg_perms file.
> 
> I'm sadly not sure how to better phrase this in the commit message, if that's still unclear please let me know and/or point me to the commit message part that's not clear enough. :)

It's not the words, but my lack of understanding of rev-parse.

After reading your comment, and as you say and by the first "cd $(dirname $0)" it assumes that the directory that we want is /etc/ngcp-config/.git (if we're not in "our config dir" the whole .ngcpcfg_perms doesn't make sense, really)... why not use that directly, instead of the dance with git-rev-parse?

Something like:
  cd ${NGCPCFG_MAIN:-/etc/ngcp-config}/.git"

But I'm fine to keep more flexibility and to fix this issue ASAP without spending more time, so maybe better just submit this as-is.

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2