TT#106202 Fix 'ngcp-config decrypt': NGCPCTL_MAIN: unbound variable
The command 'decrypt' produced an error:
> root@sp1:~# ngcp-config decrypt
> /usr/sbin/ngcp-config: line 12: NGCPCTL_MAIN: unbound variable
> root@sp1:~#

It happens due to the missing load of '"${FUNCTIONS}"/main'
inside 'scripts/decrypt' which is required for 'ngcpcfg_update_perms()'.

The '"${FUNCTIONS}"/main' cannot be loaded here as explained inside 'scripts/decrypt':
> # sadly we can't source ${FUNCTIONS}/main as we are missing a bunch of
> # configuration files that are supposed to be available, therefore
> # provide the main functions we need for successful execution of the
> # decrypt function

The fix here is to extract action 'decrypt' from ngcpcfg_update_perms()
as it is not necessary here as we are unpacking files from archive anyway.

Change-Id: Ic3117f585c2ff4af555fb7600a5b331cf7425e3c