TT#111254 Provide symlink for ngcp-config shell completion
When we invoke `ngcp-config <tab>` for the very first time in a bash
session, our shell completion doesn't work.  Once `ngcpcfg <tab>` was
invoked, then the shell completion for ngcp-config works as well though.

This is caused by the fact, that bash loads the completion file(s)
on-demand, and we register the completion for `ngcp-config` inside our
`ngcpcfg` completion file.

If we provide /usr/share/bash-completion/completions/ngcp-config as
symlink pointing to /usr/share/bash-completion/completions/ngcpcfg, then
the completion for `ngcp-config` also works on first invocation of both
ngcpcfg *and* ngcp-config.

Change-Id: I99ce535b0be6ff47c0c72ab158cc943b1a1962e3