TT#120752 Add 'ngcpcfg edit' to speedup YML editing
Normally it is necessary to press TAB multiple times, e.g:
> vim /etc/ngcp-conf<TAB>/con<TAB>f<TAB>...
For new NGCP users it is even longer.

In the same time we have a nice usability experience for ngcp-ppa,
where users are just pressing "ngcp-ppa<ENTER>1<ENTER>1<ENTER>".

Let's add the same concept here as the order of the main files is static:
> 1) /etc/ngcp-config/config.yml
> 2) /etc/ngcp-config/network.yml
> 3) /etc/ngcp-config/constants.yml

if I need to edit network.yml I will type:
> ngcpc<TAB> e<TAB><ENTER>2<ENTER>

or even:
> ngcpc<TAB> e<TAB>2<ENTER>

Change-Id: Id20fec8c931c15c73a73689a3f8b552a72ef11f7