TT#105151 When gathering ifaces for hosts, allow eth*->neth* renaming
Interfaces have to be renamed in mr9.5 for Debian bullseye, the decision
was to do eth*->neth*.

During upgrades (with the method of installing the new version in a
second partition), when we change the names in
/ngcp-fallback/etc/ngcp-config/network.yml and then "chroot
/ngcp-fallback ngcpcfg build", the config built from template
"dnsmasq.pxe_hosts" uses this function, and gets less than the expected
interfaces (only "lo", see [1]), and "dnsmasq.d/dhcp.conf" doesn't get
the IP of those "neth*" interfaces in the line "dhcp-option=" (see [2]).

This is because this function only returns interfaces which "exist", so
even if they are configured in network.yml for the new version, they do
not exist in the current/old release and live system, so they are not
returned.

One option is to remove this check altogether, since it does not make a
lot of sense.  At the time of building config, one of the interfaces
might be down, and still the configuration should be built for it.

But on the other hand, maybe there are cases like some services expect
all the configured interfaces up when they start, and they refuse to
start if this is not the case.  In general, it's changing the default of
many years, so it's a bit dangerous and can lead to obscure errors.

So the current change is a compromise: continue with the same behaviour,
but special-case this renaming of mr9.5, and if the interface in
question is named e.g. "neth0" then consider it valid if the interface
that exists is "eth0", because this is pretty much guaranteed in this
version, and in future versions we shouldn't have interfaces named
"eth0" (so they cannot be active).

[1] This happened and commit 02c737439a01382784f739882430399180d9ef93
    from templates.git ("TT#123200 Force dnsmasq listen all 'boot_int'
    interfaces from network.yml") tried to address it in another way,
    but it still fails due to what's explained above.  Presumably in the
    past it got "eth0" as the only interface to listen to, after that
    commit it gets all interfaces, but since "neth*" are marked as
    inexistent, it only gets "lo" in the end. Diff:

    --- a/dnsmasq.pxe_hosts
    +++ b/dnsmasq.pxe_hosts
    @@ -1,3 +1,3 @@
    -192.168.255.251 sp1
    +127.0.0.1 sp1
     192.168.255.252 sp2
    -192.168.255.250 sp
    + sp

[2] Diff of the file:
    -dhcp-option=6,192.168.255.251,192.168.255.252
    +dhcp-option=6,127.0.0.1,192.168.255.252

    So it gets 127.0.0.1 from "lo" interface, instead of
    "192.168.255.251" from the "neth*" one.

Change-Id: I50458dd8906899af7c7467347fdf9e1ad82a8f48