TT#160856 Improve instance's connections to support more interfaces
Defining instances connections could be necessary (for example in SEMS)
to get both the IP of the sip_int interface of LB instance and the IP
of the sip_ext interface.
This is not currently possible because we are allowing definitions of
only one interface type per connection.

The new supported structure is:
    connections:
      lb:
      - type: instance
        name: A
        interfaces:
          - name: vlan1720
            type: sip_int
          - name: vlan1721
            type: sip_ext

* pre-commit: exclude dirs from end-of-file-fixer hook

Change-Id: I832645cb916ee7f2febbea6397a439ea06fbe955