TT#149202 cat: Add new action to concatenate configuration files
This new action concatenates the various configuration files and outputs
the result as YAML. This is useful because they have a determined
ordering, and include optional host specific files that might not be
obvious how to merge.

This will also be the foundation for verifying all the current host
settings against the cfg-schema.

Change-Id: I72a61193f74caf3b2f7a58a47eefad2ed46c973a