TT#164851 Instances, add a script to fill /etc/hosts
This is a script, which properly prepares hosts records for
connections towards nosql/sql DBs.

It is able to differentiate the Carrier grade from the PRO setup.
It creates the following list of records:
- at least one record towards central node of the 'db' role (db.central);
- at least one record towards central node of the 'db' role (nosql.central) ;
- records for each PRX/LB node pointing to replicated pair and replicated central ;

In the Carrier grade:
- db.central, points to -> shared IP of the central DB (only on Proxy nodes)
- nosql.central, points to -> shared IP of the central Nosql server (only on Proxy nodes)
- LB has records (for all LB nodes) of replicated pair (nosql db)
  pointing to its 'ha_int' type IP address
- Proxy has records (for all Proxy nodes) of replicated pair (nosql db, sql db)
  and replicated central type (db only), pointing to its 'ha_int' type IP address

In the PRO setup:
- db.central, points to -> own 'ha_int' type IP address
- nosql.central, points to -> own 'ha_int' type IP address
- records (for all nodes) of replicated pair (nosql db, sql db) and
  replicated central type (sql db only), pointing to its 'ha_int' type IP address

This gives us a way for the nosql db / sql db be listening on the
internal IP addresses (not on the loopback ifaces) and not use floating IPs
to be reachable constantly by instances.

Meanwhile instances constantly remain using local nosql/sql DB
(of course excluding a case with the Carrier,
 where central sql db/nosql db has a remote shared IP).

Change-Id: Ifceb9fccab48cf78856f8aa8752c3a8b769e3f63