TT#161150 NGCP::Template::Object: Add support for spN nodes to get_nodename()
Add support for getting the nodename for any hostname in an NGCP
cluster.

We switch from fetching this information from the filesystem to the one
contained now in network.yml.

We accept up to 9 nodes per "pair", from sp1 to sp9, so for Carrier
setups that means hostnames ending from 'a' to 'i'. On CE the nodename
is called "spce".

If we cannot match the hostname to a known pattern we will error out.

Change-Id: I80c287046d02f60193479febee427615474043e3