TT#177600 Remove redundant <template>.tt2.<nodename> suffix handling
This special handling was from a time when we did not have any other
generic suffix support to override based on hostname, nodename, role,
etc. It also depends on the now obsolete systems.cfg file which is
redundant and already covered by the information contained in
network.yml.

Our tt2-process tool already handles these suffixes, so there is no
point in special-casing them from within the build script. This way we
simplify our templating and overriding mechanism, and reduce duplication
of information.

Change-Id: Iabf597dbaddf755a7f85504975afc8666a58d272