TT#139455 instances-info: Make output reproducible
We need to specify what to sort on, otherwise we are sorting based on
the references and not the actual keys. Add specific code blocks to sort
on the hash keys we are interested in. This makes the output
reproducible which makes the tests always pass, instead of randomly
failing depending on the ordering used randomly by perl.

Fixes: commit 83003649e3a5b3bde2ddebaf9db7f1b9b72da9ea
Change-Id: I3dfc031a6c76db5153b7d05f6058e88d4b101599
(cherry picked from commit e48864fc337a5b690ab7bb6ab74e53f6686b7bdc)