MT#55942 Add multi-site config loading support via sites.yml
We check whether we have both a sites.yml and a sites/ directory.
If so, and multi-sites support is enabled, then we proceed to load
all the multi-sites configurations, by placing each of them under
«sites.<site-name>.<config>». Adding a «sites.current» link to the
current site, and filling the root tree config with the current site
as well for backwards compatibility.

Change-Id: Ia810f57e8e976eef6a7582bddda46c1701456ed6