MT#56898 Fix owner and permissions for all config.*.yml files
We currently only fix up owner and permissions for known local files,
including the stock config.yml and any variant affecting the current
host (which include role selectors, pair or host specific files).

The problem is that when pushing these to the remotes where the files
are local there, then their permissions will be fixed up, and that will
happen independently for different files on different nodes. Which
results in git conflicts when trying to merge back these files.

Instead add a new variable that contains the glob for all such
customized files which we will fix up in the current node.

Change-Id: Ib2a7317a92733ad5b4d2fa8106759b00049edb68