MT#57761 Do not call restore permissions for set/del commands
In the deployment of our lab carrier, we are heavily using   ngcpcfg set
command for example to change the IP addresses of all proxies. The issue is that
execution of a command takes around 1,5 seconds and this seems because every time
"set" is used, we call the check-permissions script, which has some time
penalty to be executed. Most likely nobody uses set quite often, so that is why
the problem was not so visible for a long time.

Change-Id: I47c6c3b88b0f108aadf9d8d9a2b131e0cb76b498