MT#58014 Replace GnuPG usage with Stateless OpenPGP CLI
The usage patterns are more ergonomic, part of an IETF draft [S], and
this moves us away from GnuPG where its upstream has declared[D] it will
not be following the upcoming OpenPGP RFC [R] standards going forward.

We add support for sqop (Sequoia-PGP), gosop (GopenPGP) and
pgpainless-cli (PGPainless), but we only depend on sqop to get a
reproducible installation. This leave open the door to switch to other
implementations very easily.

As a bonus this also avoids the noise from GnuPG about creating stuff
under the user directory.

[S] https://datatracker.ietf.org/doc/draft-dkg-openpgp-stateless-cli/
[D] https://mailarchive.ietf.org/arch/msg/openpgp/Ek9SvwdsWJ_j4C7SWVoCSKwfo8Y/
[R] https://datatracker.ietf.org/doc/draft-ietf-openpgp-crypto-refresh/

Change-Id: I8a582ba1269aca3b9db6de8ddbae525c7298194c