MT#58014 encrypt, decrypt: Clarify cryptographic terms
Rename gpg to pgp or OpenPGP and remove crypted from filenames.

The standard is called OpenPGP, GnuPG (or GPG, gpg) is one of many
implementations (although depending on the context the prevalent one)
so we use that when referring to the specification.

For the encrypted tarball we remove the «-crypted» term which is an
odd wording use for encryted content and it is unnecessary and confusing
when using layered containers (as when the tarball is decrypted the
name becomes invalid); and make it use «.pgp» instead of «.gpg» as
extension, while trying the old filename on decryption for backwards
compatibility.

Change-Id: If5e0349ff0c3d8e3b47ced361e9e77d7d4d2defc