MT#59308 commit: Do not include the date in the commit message
The date is already tracked as part of the commit metadata, and we use
the current day which will always match the author date for the commit,
so no information is lost.

This makes the commit message more clear and removes redundant and
distracting information from it.

For «ngcpcfg log» we include the date from the commit metadata explicitly
in the output, which will be duplicated for old entries but gets it from
where it belongs for new ones. (Even though the date is also printed in
relative terms on the front of the line.) We now get this output for old
and new entries:

  * b3767f6 (10 hours ago) new [2024-02-02 14:28:11 +0100]  (HEAD -> master)
  * dc0a310 (3 days ago) old [2024-01-30 23:10:32.674452749+01:00] [2024-01-30 23:10:32 +0100]

Change-Id: I4eb4faed433db52e6755bc7782ec34545741d452