MT#59308 commit: Include reason in «ngcpcfg apply» commit messages
When doing an «ngcpcfg apply» we are required to pass a reason for it
which might be included in a «git commit», but not in the etckeeper
commit. Modify the wording for the latter, and include that reason so
that the git history makes more sense and is more traceable.

So when executing:

  # ngcpcfg apply "Fix array in component B"

The git commit message looked before like:

  ngcpcfg apply

And now it will look like

  Fix array in component B via "ngcpcfg apply"

Change-Id: Ib7b4ce986366ca51a487a087cb327b0d0dfe1ccd