MT#58014 decrypt: Source logs instead of open-coding its contents
These functions were previously in the main functions file, which we
could not use as it depends on configuration files we do not have
available while decrypting these same configurations. But these got
split up for performance reasons into its own functions file. We can
thus now switch to use that instead of having to keep the same functions
in sync in multiple places.

Change-Id: I3860d747e575146d4ffc1a6fc8160d7aa16925ed